Fix bundle loading problems (issue #16)
The problem: NSBundle is not able to load any framework as a bundle
(not only the JavaMemoryAdaptor) when the build is run in the console.

The activation of the NSProjectBundleEnabled property results in the
usage of an additional factory to load bundles: the NSStandardProjectBundle.Factory.
This factory is not able to create bundles for legacy *.frameworks and,
in normal circumstances, do nothing in these cases. In the WOUnitIssue16Test
project, however, the NSStandardProjectBundle.Factory creates bundles
for every *.framework mistakenly which causes the problem described.
The factory gets confused because the *.frameworks are stored inside
of the WOUnitIssue16Test project.

The solution: WOUnit now have its own NSBundleFactory that understands
the possible scenarios and delegates the bundle creation to the correct
factory.