Add support for entities that have a primary key of type Long (issue #48)
Currently, WOUnit does not support entities that have a primary key of type long. This small change fixes the issue by using a long global fake id. It takes a look at the type of the PK and casts to long/int as needed.

This fix was originally provided by Giancarlo Dessena. See pull request for more details.

https://github.com/hprange/wounit/pull/48